create and use shared metric package between controllers and pods
