+++ pkg/eraser/eraser.go
@@ -20,6 +20,7 @@ import (
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
 
 	"github.com/Azure/eraser/pkg/logger"
+	"github.com/Azure/eraser/pkg/metrics"
 
 	eraserv1alpha1 "github.com/Azure/eraser/api/v1alpha1"
 	util "github.com/Azure/eraser/pkg/utils"
@@ -141,10 +142,10 @@ func main() {
 	ctx, cancel := signal.NotifyContext(context.Background(), os.Interrupt, syscall.SIGTERM)
 	defer cancel()
 
-	exporter, reader, provider := util.ConfigureMetrics(ctx, log, os.Getenv("OTEL_EXPORTER_OTLP_ENDPOINT"))
+	exporter, reader, provider := metrics.ConfigureMetrics(ctx, log, os.Getenv("OTEL_EXPORTER_OTLP_ENDPOINT"))
 	global.SetMeterProvider(provider)
 
-	defer util.ExportMetrics(log, exporter, reader, provider)
+	defer metrics.ExportMetrics(log, exporter, reader, provider)
 
 	if err := recordMetrics(ctx); err != nil {
 		log.Error(err, "error recording metrics")
