+++ pkg/utils/utils.go
@@ -18,9 +18,6 @@ import (
 	pb "k8s.io/cri-api/pkg/apis/runtime/v1alpha2"
 
 	eraserv1alpha1 "github.com/Azure/eraser/api/v1alpha1"
-	"github.com/go-logr/logr"
-	"go.opentelemetry.io/otel/exporters/otlp/otlpmetric/otlpmetrichttp"
-	sdkmetric "go.opentelemetry.io/otel/sdk/metric"
 )
 
 const (
@@ -359,33 +356,4 @@ func WriteScanErasePipe(vulnerableImages []eraserv1alpha1.Image) error {
 	}
 
 	return file.Close()
-}
-
-func ConfigureMetrics(ctx context.Context, log logr.Logger, endpoint string) (sdkmetric.Exporter, sdkmetric.Reader, *sdkmetric.MeterProvider) {
-	exporter, err := otlpmetrichttp.New(ctx, otlpmetrichttp.WithInsecure(), otlpmetrichttp.WithEndpoint(endpoint))
-	if err != nil {
-		log.Error(err, "error initializing exporter")
-		return nil, nil, nil
-	}
-
-	reader := sdkmetric.NewPeriodicReader(exporter)
-	provider := sdkmetric.NewMeterProvider(sdkmetric.WithReader(reader))
-
-	return exporter, reader, provider
-}
-
-func ExportMetrics(log logr.Logger, exporter sdkmetric.Exporter, reader sdkmetric.Reader, provider *sdkmetric.MeterProvider) {
-	ctxB := context.Background()
-
-	m, err := reader.Collect(ctxB)
-	if err != nil {
-		log.Error(err, "failed to collect metrics")
-		return
-	}
-	if err := exporter.Export(ctxB, m); err != nil {
-		log.Error(err, "failed to export metrics")
-	}
-	if err := provider.Shutdown(ctxB); err != nil {
-		log.Error(err, "error during metric shutdown", err)
-	}
 }
